alterações na parte de experiencia, torcedor agora tem sua pontuação decrementada após aquisição de experiencia

+++ src/main/java/br/usjt/apivolei/maestro/model/interfaces/IPontoTorcedor.java
@@ -1,6 +1,6 @@
 package br.usjt.apivolei.maestro.model.interfaces;
 
 public interface IPontoTorcedor {
-    Double incrementar(Double ponto);
-    Double decrementar(Double ponto);
+    Integer incrementar(Integer ponto);
+    Integer decrementar(Integer ponto);
 }
